Split into specific tests and refactoring common code
